backend/services/calendar-service/requirements.txt
fastapi==0.115.0
uvicorn==0.30.6
databases[aiosqlite]==0.9.0
aiosqlite==0.20.0
pydantic==2.9.1
python-dotenv==1.0.1
httpx==0.27.2
google-auth==2.35.0
google-auth-oauthlib==1.2.1
google-api-python-client==2.149.0
python-jose[cryptography]==3.3.0
